requirements.txt
python-telegram-bot==20.8
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.9.0